Commit aula03 JavaAdvanced Fiap

+++ src/main/java/br/com/fiap/api_rest/controller/LivroController.java
@@ -10,64 +10,60 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @RestController
 @RequestMapping(value = "/livros")
 public class LivroController {
+
     @Autowired
     private LivroRepository livroRepository;
+
     @Autowired
     private LivroService livroService;
 
-    // CREATE, READ, UPDATE, DELETE
-    // POST, GET, PUT, DELETE
-
     @PostMapping
-    public ResponseEntity<Livro> createLivro(@RequestBody LivroRequest livro) {
-        Livro livroSalvo = livroRepository.save(livroService.requestToLivro(livro));
-        return new ResponseEntity<>(livroSalvo,HttpStatus.CREATED);
+    public ResponseEntity<Livro> createLivro(@RequestBody LivroRequest livroRequest) {
+        Livro livroSalvo = livroRepository.save(livroService.requestToLivro(livroRequest));
+        return new ResponseEntity<>(livroSalvo, HttpStatus.CREATED);
     }
 
     @GetMapping
     public ResponseEntity<List<LivroResponse>> readLivros() {
         List<Livro> livros = livroRepository.findAll();
-        List<LivroResponse> listaLivros = new ArrayList<>();
-        for (Livro livro : livros) {
-            listaLivros.add(livroService.livroToResponse(livro));
-        }
-        return new ResponseEntity<>(listaLivros,HttpStatus.OK);
+        return new ResponseEntity<>(livroService.livrosToResponse(livros), HttpStatus.OK);
     }
 
-    // @PathVariable localhost:8080/livros/1
-    // @RequestParam localhost:8080/livros/?id=1
     @GetMapping("/{id}")
-    public ResponseEntity<Livro> readLivro(@PathVariable Long id) {
+    public ResponseEntity<LivroResponse> readLivro(@PathVariable Long id) {
         Optional<Livro> livro = livroRepository.findById(id);
         if (livro.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(livro.get(),HttpStatus.OK);
+        return new ResponseEntity<>(livroService.livroToResponse(livro.get()), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Livro> updateLivro(@RequestBody Livro livro,
+    public ResponseEntity<Livro> updateLivro(@RequestBody LivroRequest livroRequest,
                                              @PathVariable Long id) {
         Optional<Livro> livroExistente = livroRepository.findById(id);
         if (livroExistente.isEmpty()) {
             return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
         }
-        livro.setId(livroExistente.get().getId());
-        Livro livroSalvo = livroRepository.save(livro);
-        return new ResponseEntity<>(livroSalvo,HttpStatus.CREATED);
+        Livro livroAtualizado = livroService.requestToLivro(livroRequest);
+        livroAtualizado.setId(id);
+        Livro livroSalvo = livroRepository.save(livroAtualizado);
+        return new ResponseEntity<>(livroSalvo, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteLivro(@PathVariable Long id) {
+        Optional<Livro> livro = livroRepository.findById(id);
+        if (livro.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         livroRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
 }
